refactor(player): rename components and contexts to youtube
This commit renames all player-related components and contexts to reflect the use of the Youtube player.
It updates import paths and component names to align with the new naming convention.

The changes primarily involve renaming files and updating references throughout the project.
